20220818 web sales

+++ SallesWebMvc/SallesWebMvc/Models/SellesRecord.cs
@@ -6,7 +6,7 @@ namespace SallesWebMvc.Models
     {
         public int Id { get; set; }
         public DateTime Date { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public SalleStatus Status { get; set; }
         public Seller Seller { get; set; }
 
@@ -14,7 +14,7 @@ namespace SallesWebMvc.Models
         {
         }
 
-        public SellesRecord(int id, DateTime date, double amount, SalleStatus status, Seller seller)
+        public SellesRecord(int id, DateTime date, decimal amount, SalleStatus status, Seller seller)
         {
             Id = id;
             Date = date;
